requirements.txt
Django>=2.0,<3.0
mysqlclient==1.4.6
django-import-export==2.0.1
djangorestframework==3.11.0